infra/kpaas/pyproject.toml
[tool.poetry]
name = "kpaas"
version = "0.1.0"
description = ""
authors = ["voidfiles <voidfiles@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
annotated-types = "0.7.0"
ansible = "10.1.0"
ansible-core = "2.17.1"
anyio = "4.4.0"
asyncache = "0.3.1"
boto3 = "1.34.134"
botocore = "1.34.134"
cachetools = "5.3.3"
certifi = "2024.6.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cryptography = "42.0.8"
dnspython = "2.6.1"
exceptiongroup = "1.2.1"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
httpx-ws = "0.6.0"
idna = "3.7"
jinja2 = "3.1.4"
jmespath = "1.0.1"
kr8s = "0.16.1"
markupsafe = "2.1.5"
packaging = "24.1"
pycparser = "2.22"
pydantic = "2.7.4"
pydantic-core = "2.18.4"
pyhelm3 = "0.3.3"
python-box = "7.2.0"
python-dateutil = "2.9.0.post0"
python-jsonpath = "1.1.1"
pyyaml = "6.0.1"
requests = "2.32.3"
resolvelib = "1.0.1"
s3transfer = "0.10.2"
six = "1.16.0"
sniffio = "1.3.1"
typing-extensions = "4.12.2"
urllib3 = "2.2.2"
wsproto = "1.2.0"
typer = "^0.12.3"
dataclass-wizard = "^0.22.3"
pyonepassword = "^5.0.0"

[tool.poetry.scripts]
kpaas = "kpaas.cmd:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
